Update MonoDevelop.PackageManagement files and NuGet dlls to release-5.8
NuGet.Client release-5.8.x (and later) is able to handle .net5.0 projects
To allow DotDevelop to build with this release of Nuget a number of small changes
are needed to the DotDevelop PackageManagement files because the interface to
NuGet changed after release-5.4.0 (which DotDevelop currently builds with)

The commits in this 'net5ff_nuget' branch also remove 'nuget-binary' from being
an external submodule within the DotDevelop repo, since 'nuget-binary' has not
been updated since MonoDevelop was archived. Instead, the folder with the NuGet dlls
is now a simple folder under DotDevelop's version control. This simplifies updating
NuGet to later releases, which will be required to handle .net6.0+ frameworks.

With these commits, net5.0 projects can be created, saved, etc, which effectively fixes
Issue #79 (project.assets.json gets a wrong TargetFrameworkMoniker for net5 and higher)